code/core/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-ninja==1.1.0
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
pydotplus==2.0.2
PyJWT==2.8.0
pyparsing==3.1.2
python-dotenv==1.0.1
python-ipware==3.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
